Updated change log entry.
Removed __doc__ from whitelisted internals and changed ai.cpp to
remove signed/unsigned compiler warnings.

+++ src/ai.cpp
@@ -221,7 +221,7 @@ ai_interface* create_ai(const std::string& name, ai_interface::info& info)
 		return new ai2(info);
 	else if(name == "python_ai")
 #ifdef HAVE_PYTHON
-		return new python_ai(info);
+	  return new python_ai(info);
 #else
     {
 		LOG_STREAM(err, ai) << "No Python AI support available in this Wesnoth build!\n";
@@ -2239,14 +2239,14 @@ variant ai_interface::get_value(const std::string& key) const
 		return variant(get_info().state.turn());
 	} else if(key == "allies") {
 		std::vector<variant> vars;
-		for( int i = 0; i < info_.teams.size(); ++i) {
+		for( size_t i = 0; i < info_.teams.size(); ++i) {
 			if ( !info_.teams[info_.team_num-1].is_enemy( i+1 ) )
 				vars.push_back(variant( i ));
 		}
 		return variant(&vars);
 	} else if(key == "enemies") {
 		std::vector<variant> vars;
-		for( int i = 0; i < info_.teams.size(); ++i) {
+		for( size_t i = 0; i < info_.teams.size(); ++i) {
 			if ( info_.teams[info_.team_num-1].is_enemy( i+1 ) )
 				vars.push_back(variant( i ));
 		}
@@ -2260,7 +2260,7 @@ variant ai_interface::get_value(const std::string& key) const
 	} else if(key == "units_of_team") {
 		std::vector<variant> vars;
 		std::vector< std::vector< variant> > tmp;
-		for( int i = 0; i<info_.teams.size(); ++i)
+		for( size_t i = 0; i<info_.teams.size(); ++i)
 		{
 			std::vector<variant> v;
 			tmp.push_back( v );
@@ -2268,7 +2268,7 @@ variant ai_interface::get_value(const std::string& key) const
 		for(unit_map::const_iterator i = info_.units.begin(); i != info_.units.end(); ++i) {
 			tmp[ i->second.side()-1 ].push_back( variant(new unit_callable(*i)) );
 		}
-		for( int i = 0; i<tmp.size(); ++i)
+		for( size_t i = 0; i<tmp.size(); ++i)
 			vars.push_back( variant( &tmp[i] ));
 		return variant(&vars);
 	} else if(key == "my_units") {
